BaseOperation -> BaseViewModel
